cms_django/requirements.txt
anyio==4.10.0
asgiref==3.9.1
certifi==2025.8.3
channels==4.1.0
click==8.2.1
Django==5.2.5
django-environ==0.11.2
djangorestframework==3.16.1
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
idna==3.10
mysqlclient==2.2.7
python-dotenv==1.1.1
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.14.1
uvicorn==0.30.6
